URI ONLINE JUDGE

+++ 01 - Iniciante/0096 - Preenchimento de Vetor III.py
@@ -1,4 +1,4 @@
-x = int(input())
+x = float(input())
 n = []
 n.append(x)
